refactor: Ajout de libs/shared pour optimiser le point de terminaison de lecture et refactoring du consume event de ProductRegistryCommandRessource
